requirements.txt
appnope==0.1.0
arrow==0.12.1
asn1crypto==0.24.0
atomicwrites==1.1.5
attrs==18.1.0
aws==0.2.5
aws-xray-sdk==0.95
backcall==0.1.0
backports.functools-lru-cache==1.5
backports.ssl-match-hostname==3.5.0.1
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==3.1.4
binaryornot==0.4.4
boto==2.49.0
boto3==1.7.71
botocore==1.10.71
certifi==2019.11.28
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
configparser==3.5.0
cookiecutter==1.6.0
cookies==2.2.1
coverage==4.5.1
coveralls==1.3.0
cryptography==2.3
decorator==4.3.0
docker==3.4.1
docker-pycreds==0.3.0
dockerflow==2018.4.0
docopt==0.6.2
docutils==0.14
entrypoints==0.3
enum34==1.1.6
fabric==2.1.3
flake8==3.7.7
Flask==1.0.2
Flask-API==1.0
funcsigs==1.0.2
future==0.16.0
idna==2.7
importlib-metadata==1.5.0
invoke==1.1.0
ipaddress==1.0.22
iso8601==0.1.12
itsdangerous==0.24
jedi==0.12.1
Jinja2==2.10.1
jinja2-time==0.2.0
jmespath==0.9.3
jsondiff==1.1.1
jsonpickle==0.9.6
MarkupSafe==1.0
mccabe==0.6.1
mock==2.0.0
more-itertools==4.2.0
moto==1.3.3
mozilla-srgutil==0.2.1
numpy==1.14.3
packaging==17.1
paramiko==2.4.2
parso==0.3.1
pathlib2==2.3.2
pbr==4.1.0
pexpect==4.6.0
pickleshare==0.7.4
pip-api==0.0.1
pkginfo==1.4.2
pluggy==0.13.1
ply==3.11
poyo==0.4.1
prettytable==0.7.2
prompt-toolkit==1.0.15
ptyprocess==0.6.0
py==1.5.3
pyaml==17.12.1
pyasn1==0.4.3
pycodestyle==2.5.0
pycparser==2.18
pyflakes==2.1.1
Pygments==2.2.0
PyNaCl==1.2.1
pyparsing==2.2.0
pytest==5.3.5
pytest-cov==2.5.1
pytest-flask==0.14.0
python-dateutil==2.6.1
python-decouple==3.1
pytz==2018.5
PyYAML==4.2b1
requests==2.20.1
requests-toolbelt==0.8.0
responses==0.9.0
rsa==3.4.2
s3transfer==0.1.13
scandir==1.8
scipy==1.1.0
sentry-sdk==0.7.3
setuptools-scm==2.1.0
simplegeneric==0.8.1
six==1.11.0
spark==0.2.1
spark-parser==1.8.7
tox==3.0.0
tqdm==4.23.4
translationstring==1.3
twine==1.11.0
uncompyle2==2.0.0
uncompyle6==3.2.0
urllib3==1.25
virtualenv==16.0.0
wcwidth==0.1.7
websocket-client==0.48.0
Werkzeug==0.16.1
whichcraft==0.4.1
wrapt==1.10.11
xdis==3.8.2
xmltodict==0.11.0
zipp==3.0.0
